+++ app/src/main/java/eu/faircode/email/DB.java
@@ -68,7 +68,7 @@ import io.requery.android.database.sqlite.SQLiteDatabase;
 // https://developer.android.com/topic/libraries/architecture/room.html
 
 @Database(
-        version = 215,
+        version = 216,
         entities = {
                 EntityIdentity.class,
                 EntityAccount.class,
@@ -416,7 +416,7 @@ public abstract class DB extends RoomDatabase {
                         // https://www.sqlite.org/pragma.html
                         for (String pragma : new String[]{
                                 "synchronous", "journal_mode",
-                                "wal_checkpoint", "wal_autocheckpoint",
+                                "wal_checkpoint", "wal_autocheckpoint", "journal_size_limit",
                                 "page_count", "page_size", "max_page_count", "freelist_count",
                                 "cache_size", "cache_spill",
                                 "soft_heap_limit", "hard_heap_limit", "mmap_size",
@@ -2191,6 +2191,12 @@ public abstract class DB extends RoomDatabase {
                     public void migrate(@NonNull SupportSQLiteDatabase db) {
                         Log.i("DB migration from version " + startVersion + " to " + endVersion);
                         db.execSQL("ALTER TABLE `folder` ADD COLUMN `auto_add` INTEGER");
+                    }
+                }).addMigrations(new Migration(215, 216) {
+                    @Override
+                    public void migrate(@NonNull SupportSQLiteDatabase db) {
+                        Log.i("DB migration from version " + startVersion + " to " + endVersion);
+                        db.execSQL("ALTER TABLE `message` ADD COLUMN `infrastructure` TEXT");
                     }
                 }).addMigrations(new Migration(998, 999) {
                     @Override