+++ app/src/main/java/eu/faircode/email/ViewCardOptional.java
@@ -33,6 +33,7 @@ import androidx.preference.PreferenceManager;
 public class ViewCardOptional extends CardView {
     private boolean cards;
     private boolean compact;
+    private int padding;
     private int margin;
     private int ident;
     private Integer color = null;
@@ -56,10 +57,12 @@ public class ViewCardOptional extends CardView {
         SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
         cards = prefs.getBoolean("cards", true);
         compact = prefs.getBoolean("compact", false);
+        padding = prefs.getInt("view_padding", compact ? 0 : 1);
 
         // dev4223: was: '3 : 6' and '12 + (compact ? 3 : 6)'
-        margin = Helper.dp2pixels(context, compact ? 1 : 2);
-        ident = Helper.dp2pixels(context, 6 + (compact ? 2 : 4));
+        // old: margin = Helper.dp2pixels(context, compact ? 1 : 2);
+        // old: ident = Helper.dp2pixels(context, 6 + (compact ? 2 : 4));
+        margin = Helper.dp2pixels(context, (padding + 1) * 3);
 
         // dev4223: was: 'vsetRadius(cards ? margin : 0);'
         setRadius(cards ? 4 : 0);